Version 1.4.0.4
build_dialogs.py: This file auto magically takes a built in dictionary and assembles dialogs.hpp. It does a good job sorting the buttons, but an even better job of simplifying the task of adding new ones or reorganizing old ones.

initVar.sqf: This is a temporary change to stop the swarms of helicoptors that come crashing into AA nests with no real effect on the war.... basically this forces a land battle. (Looking to rework air attack chances past a simple to close to far check)

resourcecheck: added a tax system for multiplayer and server member enabled games. Essentially takes a hard coded 5% of earnings and distributed it evenly to the players. This helps them keep enough money in their personal coffers to have enough cash to buy an AI or two and transportation.

dialogs: this is an automatically generated file now.
References: This is where key pieces of information are kept in order to properly update the templates with addtional assets for now its just RHS asset lists.

build.bat: This simplifies the process of building and copying the pbo where it belongs. (Run, Manually pack PBO, run again to put pbo in test folder.)